requirements.txt
Django==5.0.5
djangorestframework==3.15.1